+++ srcs/parse_unsigned.c
@@ -1,7 +1,7 @@
 #include "../includes/ft_printf.h"
 
 static void get_result_str(t_sides side, unsigned int number,
-						   int num_len, char *result)
+						   size_t num_len, char *result)
 {
 	char *itoa_str;
 
@@ -22,12 +22,12 @@ static char *str_from_arg(t_parameter parameter, unsigned int number)
 {
 	t_sides sides;
 	char *result;
-	int max_len;
-	int num_len;
+	size_t max_len;
+	size_t num_len;
 
 	num_len = get_unsigned_num_len(number);
-	if (num_len >= parameter.num_before_dot
-		&& num_len >= parameter.num_after_dot)
+	if ((int)num_len >= parameter.num_before_dot
+		&& (int)num_len >= parameter.num_after_dot)
 		return (ft_unsigned_itoa(number));
 	if (parameter.num_before_dot > parameter.num_after_dot)
 		max_len = parameter.num_before_dot;
